// app/types/employee.ts
export interface Employee {
  sno: string;
  manglaTimePicture?: string;
  currentPicture?: string;
  name: string;
  phoneNumber: string;
  servedAs: string;
  workedInSection: string;
  servedFromTo: string;
  residenceDuringJob: string;
  presentResidence: string;
  presentStatus: string;
  imageUrl?: string;
}